refactor: move RangeLookupError to standalone file. add lib cele as a container of types used by Server
