+++ src/lib.rs
@@ -1,0 +1,6 @@
+#[macro_use]
+extern crate quick_error;
+
+mod errors;
+pub use errors::*;
+
